created ducks structure
