+++ src/redux/ducks/index.js
@@ -1,0 +1,10 @@
+import employers from './employers';
+import searchedName from './searchedName';
+import subordinates from './subordinates';
+import {combineReducers} from 'redux';
+
+export default combineReducers({
+    employers:employers,
+    searchedName:searchedName,
+    subordinates:subordinates
+});
