+++ src/redux/ducks/searchedName.js
@@ -1,0 +1,23 @@
+
+export const SET_SEARCHED_NAME = 'SET_SEARCHED_NAME';
+
+const initialState = {
+    searchedName: ''
+};
+
+const reducer = (state = initialState, action) => {
+
+    switch (action.type) {
+
+        case SET_SEARCHED_NAME:
+            return {...state, searchedName: action.searchedName};
+        default:
+            return state;
+    }
+
+}
+
+export const setSearchedName = searchedName => ({type: SET_SEARCHED_NAME, searchedName})
+
+
+export default reducer;
